feat: UI - 설정화면 조작버튼 UI 개발, 게임 조작을 위해 사용될 키를 설정
- 키 설정 버튼 입력 시 현재 설정되어 있는 키 보여줌
- 텍스트 클릭 후 사용자가 원하는 임의의 버튼으로 키설정 변경가능
- "기본 설정으로 되돌리기" 버튼 입력 시 초기 설정으로 돌아감.
[!] 현재 중복된 키 입력 시 그대로 설정해버리는 버그 있음. 해당 사항 수정하기 위해 지속적으로 노력중.
